+++ secator/runners/_base.py
@@ -96,6 +96,7 @@ class Runner:
 		self.celery_ids_map = {}
 		self.caller = self.run_opts.get('caller', None)
 		self.threads = []
+		self.no_poll = self.run_opts.get('no_poll', False)
 
 		# Determine exporters
 		exporters_str = self.run_opts.get('output') or self.default_exporters
@@ -139,7 +140,7 @@ class Runner:
 		self.print_progress = self.run_opts.get('print_progress', False) and not self.quiet and not self.print_raw
 		self.print_target = self.run_opts.get('print_target', False) and not self.quiet and not self.print_raw
 		self.print_stat = self.run_opts.get('print_stat', False) and not self.quiet and not self.print_raw
-		self.raise_on_error = self.run_opts.get('raise_on_error', not self.sync)
+		self.raise_on_error = self.run_opts.get('raise_on_error', False)
 		self.print_opts = {k: v for k, v in self.__dict__.items() if k.startswith('print_') if v}
 
 		# Debug
@@ -635,6 +636,8 @@ class Runner:
 
 	def log_results(self):
 		"""Log runner results."""
+		if self.no_poll:
+			return
 		self.done = True
 		self.progress = 100
 		self.end_time = datetime.fromtimestamp(time())