+++ secator/runners/workflow.py
@@ -26,6 +26,7 @@ class Workflow(Runner):
 		# Task opts
 		run_opts = self.run_opts.copy()
 		run_opts['hooks'] = self._hooks.get(Task, {})
+		run_opts.pop('no_poll', False)
 
 		# Build Celery workflow
 		workflow = self.build_celery_workflow(
@@ -46,6 +47,8 @@ class Workflow(Runner):
 				message=f'Celery task created: {self.celery_result.id}',
 				task_id=self.celery_result.id
 			)
+			if self.no_poll:
+				return
 			results = CeleryData.iter_results(
 				self.celery_result,
 				ids_map=self.celery_ids_map,
@@ -70,7 +73,7 @@ class Workflow(Runner):
 			self.inputs,
 			self.config.options,
 			run_opts)
-		sigs = [forward_results.si(results).set(queue='io')] + sigs + [forward_results.s().set(queue='io')]
+		sigs = [forward_results.si(results).set(queue='results')] + sigs + [forward_results.s().set(queue='results')]
 		workflow = chain(*sigs)
 		return workflow
 
@@ -102,7 +105,7 @@ class Workflow(Runner):
 					workflow_opts,
 					run_opts
 				)
-				sig = chord((tasks), forward_results.s().set(queue='io'))
+				sig = chord((tasks), forward_results.s().set(queue='results'))
 			elif task_name == '_chain':
 				tasks = self.get_tasks(
 					task_opts,